refactor: 테스트 mock 객체 수정
