+++ melly-core/src/main/resources/application-test.yml
@@ -1,7 +1,4 @@
 spring:
-  mongodb:
-    embedded:
-      version: 3.5.5
   redis:
     host: 127.0.0.1
     port: 6379
