+++ melly-core/src/test/java/cmc/mellyserver/mellycore/common/db/DatabaseCleanup.java
@@ -1,8 +1,8 @@
 package cmc.mellyserver.mellycore.common.db;
 
 import com.google.common.base.CaseFormat;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.PostConstruct;
@@ -13,7 +13,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Component
-@ActiveProfiles("test")
+@Profile("test")
 public class DatabaseCleanup {
 
     @PersistenceContext
