+++ melly-core/src/test/java/cmc/mellyserver/mellycore/common/db/RedisTestContainers.java
@@ -2,24 +2,23 @@ package cmc.mellyserver.mellycore.common.db;
 
 import org.junit.jupiter.api.DisplayName;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.utility.DockerImageName;
 
 @DisplayName("Redis Test Containers")
 @Configuration
+@Profile("test")
 public class RedisTestContainers {
 
     private static final String REDIS_DOCKER_IMAGE = "redis:5.0.3-alpine"; // redis 도커 이미지 명시
 
-    static {    // (1)
-        GenericContainer<?> REDIS_CONTAINER =
-                new GenericContainer<>(DockerImageName.parse(REDIS_DOCKER_IMAGE))
-                        .withExposedPorts(6379)
-                        .withReuse(true);
+    static {
+        GenericContainer<?> REDIS_CONTAINER = new GenericContainer<>(DockerImageName.parse(REDIS_DOCKER_IMAGE))
+                .withExposedPorts(6379).withReuse(true);
 
-        REDIS_CONTAINER.start();    // (2)
+        REDIS_CONTAINER.start();
 
-        // (3)
         System.setProperty("spring.redis.host", REDIS_CONTAINER.getHost());
         System.setProperty("spring.redis.port", REDIS_CONTAINER.getMappedPort(6379).toString());
     }
