+++ melly-core/src/test/java/cmc/mellyserver/mellycore/common/mock/MockMailCertificationService.java
@@ -1,10 +1,12 @@
-package cmc.mellyserver.mellycore.config.mockapi;
+package cmc.mellyserver.mellycore.common.mock;
 
 import cmc.mellyserver.mellycore.infrastructure.email.certification.EmailCertificationRequest;
 import cmc.mellyserver.mellycore.infrastructure.email.certification.EmailCertificationService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
+@Slf4j
 @Component
 @Profile("test")
 public class MockMailCertificationService implements EmailCertificationService {
