+++ melly-core/src/test/java/cmc/mellyserver/mellycore/common/mock/MockMailService.java
@@ -1,11 +1,13 @@
-package cmc.mellyserver.mellycore.config.mockapi;
+package cmc.mellyserver.mellycore.common.mock;
 
 import cmc.mellyserver.mellycore.infrastructure.email.EmailSendService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
 
+@Slf4j
 @Component
 @Profile("test")
 public class MockMailService implements EmailSendService {
